requirements.txt
accelerate==1.9.0
aiofiles==24.1.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.8.1
asttokens==3.0.1
async-lru==2.2.0
attrs==25.3.0
audioop-lts==0.2.2
babel==2.18.0
barcode==1.0.4
bcrypt==5.0.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
Brotli==1.1.0
build==1.4.4
CacheControl==0.14.3
cachetools==5.5.2
certifi==2026.2.25
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.5.8
click==8.2.1
comm==0.2.3
contourpy==1.3.2
cryptography==45.0.5
cssselect2==0.8.0
curl_cffi==0.13.0
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.8.0
durationpy==0.10
email-validator==2.3.0
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.135.2
fastapi-cli==0.0.24
fastapi-cloud-cli==0.17.0
fastar==0.11.0
fastjsonschema==2.21.2
ffmpy==1.0.0
filelock==3.18.0
firebase-admin==6.9.0
flatbuffers==25.12.19
fonttools==4.58.4
fqdn==1.5.1
frozendict==2.4.7
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-genai==1.24.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio==6.10.0
gradio_client==2.4.0
groovy==0.1.2
groq==1.1.2
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.2.0
hf-gradio==0.3.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.2
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.1
importlib_resources==7.1.0
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kiwisolver==1.4.8
kubernetes==35.0.0
lark==1.3.1
Levenshtein==0.27.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
mmh3==5.2.1
mpmath==1.3.0
msgpack==1.1.1
multitasking==0.0.12
narwhals==1.47.1
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.5.4
notebook_shim==0.2.4
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.25.0
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.41.1
opentelemetry-exporter-otlp-proto-common==1.41.1
opentelemetry-exporter-otlp-proto-grpc==1.41.1
opentelemetry-proto==1.41.1
opentelemetry-sdk==1.41.1
opentelemetry-semantic-conventions==0.62b1
orjson==3.11.7
overrides==7.7.0
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.6
pdf2image==1.17.0
pdfkit==1.0.0
peewee==4.0.3
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.9.2
prometheus_client==0.24.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.11.1
pydantic-settings==2.14.0
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.9.0
pyparsing==3.2.3
pypdf==6.9.2
PyPDF2==3.0.1
pyphen==0.17.2
PyPika==0.51.1
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.0.0
python-Levenshtein==0.27.1
python-multipart==0.0.22
python-package-size==1.0.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
qrcode==8.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
rpds-py==0.26.0
rsa==4.9.1
safehttpx==0.1.7
safetensors==0.5.3
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==2.1.0
sentry-sdk==2.58.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.52.1
streamlit==1.47.0
sympy==1.14.0
tenacity==8.5.0
terminado==0.18.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.2
toml==0.10.2
tomlkit==0.13.3
torch==2.11.0
torchvision==0.26.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.14.0
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.6.0
weasyprint==66.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
yfinance==1.2.0
zipp==3.23.1
zopfli==0.2.3.post1
